+++ spec/spec_helper.rb
@@ -5,6 +5,7 @@ Coveralls.wear!
 
 RSpec.configure do |config|
   config.before { allow($stdout).to receive(:puts) }
+  config.order = "random"
   config.mock_with :rspec do |c|
     c.syntax = [:should, :expect]
   end